src/requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
django-phonenumber-field==5.1.0
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Markdown==3.3.4
phonenumbers==8.12.21
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
